+++ falcon/request.py
@@ -51,9 +51,9 @@ SimpleCookie = http_cookies.SimpleCookie
 DEFAULT_ERROR_LOG_FORMAT = (u'{0:%Y-%m-%d %H:%M:%S} [FALCON] [ERROR]'
                             u' {1} {2}{3} => ')
 
-TRUE_STRINGS = ('true', 'True', 'yes', '1', 'on')
-FALSE_STRINGS = ('false', 'False', 'no', '0', 'off')
-WSGI_CONTENT_HEADERS = ('CONTENT_TYPE', 'CONTENT_LENGTH')
+TRUE_STRINGS = frozenset(['true', 'True', 'yes', '1', 'on'])
+FALSE_STRINGS = frozenset(['false', 'False', 'no', '0', 'off'])
+WSGI_CONTENT_HEADERS = frozenset(['CONTENT_TYPE', 'CONTENT_LENGTH'])
 
 # PERF(kgriffs): Avoid an extra namespace lookup when using these functions
 strptime = datetime.strptime
@@ -453,8 +453,8 @@ class Request(object):
             if self.query_string:
                 self._params = parse_query_string(
                     self.query_string,
-                    keep_blank_qs_values=self.options.keep_blank_qs_values,
-                    parse_qs_csv=self.options.auto_parse_qs_csv,
+                    keep_blank=self.options.keep_blank_qs_values,
+                    csv=self.options.auto_parse_qs_csv,
                 )
 
             else:
@@ -662,7 +662,7 @@ class Request(object):
         except ValueError:
             href = 'http://goo.gl/zZ6Ey'
             href_text = 'HTTP/1.1 Range Requests'
-            msg = ('It must be a range formatted according to RFC 7233.')
+            msg = 'It must be a range formatted according to RFC 7233.'
             raise errors.HTTPInvalidHeader(msg, 'Range', href=href,
                                            href_text=href_text)
 
@@ -1164,8 +1164,8 @@ class Request(object):
 
         raise errors.HTTPMissingParam(name)
 
-    def get_param_as_int(self, name, required=False, min=None,
-                         max=None, store=None, default=None):
+    def get_param_as_int(self, name, required=False, min_value=None,
+                         max_value=None, store=None, default=None):
         """Return the value of a query string parameter as an int.
 
         Args:
@@ -1176,12 +1176,12 @@ class Request(object):
                 ``HTTPBadRequest`` instead of returning ``None`` when the
                 parameter is not found or is not an integer (default
                 ``False``).
-            min (int): Set to the minimum value allowed for this
-                param. If the param is found and it is less than min, an
+            min_value (int): Set to the minimum value allowed for this
+                param. If the param is found and it is less than min_value, an
                 ``HTTPError`` is raised.
-            max (int): Set to the maximum value allowed for this
+            max_value (int): Set to the maximum value allowed for this
                 param. If the param is found and its value is greater than
-                max, an ``HTTPError`` is raised.
+                max_value, an ``HTTPError`` is raised.
             store (dict): A ``dict``-like object in which to place
                 the value of the param, but only if the param is found
                 (default ``None``).
@@ -1198,7 +1198,7 @@ class Request(object):
                 it was required to be there, or it was found but could not
                 be converted to an ``int``. Also raised if the param's value
                 falls outside the given interval, i.e., the value must be in
-                the interval: min <= value <= max to avoid triggering an error.
+                the interval: min_value <= value <= max_value to avoid triggering an error.
 
         """
 
@@ -1217,12 +1217,12 @@ class Request(object):
                 msg = 'The value must be an integer.'
                 raise errors.HTTPInvalidParam(msg, name)
 
-            if min is not None and val < min:
-                msg = 'The value must be at least ' + str(min)
+            if min_value is not None and val < min_value:
+                msg = 'The value must be at least ' + str(min_value)
                 raise errors.HTTPInvalidParam(msg, name)
 
-            if max is not None and max < val:
-                msg = 'The value may not exceed ' + str(max)
+            if max_value is not None and max_value < val:
+                msg = 'The value may not exceed ' + str(max_value)
                 raise errors.HTTPInvalidParam(msg, name)
 
             if store is not None:
@@ -1235,8 +1235,8 @@ class Request(object):
 
         raise errors.HTTPMissingParam(name)
 
-    def get_param_as_float(self, name, required=False, min=None,
-                           max=None, store=None, default=None):
+    def get_param_as_float(self, name, required=False, min_value=None,
+                           max_value=None, store=None, default=None):
         """Return the value of a query string parameter as an float.
 
         Args:
@@ -1247,12 +1247,12 @@ class Request(object):
                 ``HTTPBadRequest`` instead of returning ``None`` when the
                 parameter is not found or is not an float (default
                 ``False``).
-            min (float): Set to the minimum value allowed for this
-                param. If the param is found and it is less than min, an
+            min_value (float): Set to the minimum value allowed for this
+                param. If the param is found and it is less than min_value, an
                 ``HTTPError`` is raised.
-            max (float): Set to the maximum value allowed for this
+            max_value (float): Set to the maximum value allowed for this
                 param. If the param is found and its value is greater than
-                max, an ``HTTPError`` is raised.
+                max_value, an ``HTTPError`` is raised.
             store (dict): A ``dict``-like object in which to place
                 the value of the param, but only if the param is found
                 (default ``None``).
@@ -1269,7 +1269,7 @@ class Request(object):
                 it was required to be there, or it was found but could not
                 be converted to an ``float``. Also raised if the param's value
                 falls outside the given interval, i.e., the value must be in
-                the interval: min <= value <= max to avoid triggering an error.
+                the interval: min_value <= value <= max_value to avoid triggering an error.
 
         """
 
@@ -1288,12 +1288,12 @@ class Request(object):
                 msg = 'The value must be a float.'
                 raise errors.HTTPInvalidParam(msg, name)
 
-            if min is not None and val < min:
-                msg = 'The value must be at least ' + str(min)
+            if min_value is not None and val < min_value:
+                msg = 'The value must be at least ' + str(min_value)
                 raise errors.HTTPInvalidParam(msg, name)
 
-            if max is not None and max < val:
-                msg = 'The value may not exceed ' + str(max)
+            if max_value is not None and max_value < val:
+                msg = 'The value may not exceed ' + str(max_value)
                 raise errors.HTTPInvalidParam(msg, name)
 
             if store is not None:
@@ -1373,8 +1373,14 @@ class Request(object):
         raise errors.HTTPMissingParam(name)
 
     def get_param_as_bool(self, name, required=False, store=None,
-                          blank_as_true=False, default=None):
-        """Return the value of a query string parameter as a boolean
+                          blank_as_true=True, default=None):
+        """Return the value of a query string parameter as a boolean.
+
+        This method treats valueless parameters as flags. By default, if no
+        value is provided for the parameter in the query string, ``True`` is
+        assumed and returned. If the parameter is missing altogether, ``None``
+        is returned as with other ``get_param_*()`` methods, which can be
+        easily treated as falsy by the caller as needed.
 
         The following boolean strings are supported::
 
@@ -1392,19 +1398,19 @@ class Request(object):
             store (dict): A ``dict``-like object in which to place
                 the value of the param, but only if the param is found (default
                 ``None``).
-            blank_as_true (bool): If ``True``, an empty string value will be
-                treated as ``True`` (default ``False``). Normally empty strings
-                are ignored; if you would like to recognize such parameters, you
-                must set the `keep_blank_qs_values` request option to ``True``.
-                Request options are set globally for each instance of
-                ``falcon.API`` through the `req_options` attribute.
-            default (any): If the param is not found returns the
-                given value instead of ``None``
+            blank_as_true (bool): Valueless query string parameters
+                are treated as flags, resulting in ``True`` being
+                returned when such a parameter is present, and ``False``
+                otherwise. To require the client to explicitly opt-in to a
+                truthy value, pass ``blank_as_true=False`` to return ``False``
+                when a value is not specified in the query string.
+            default (any): If the param is not found, return this
+                value instead of ``None``.
 
         Returns:
             bool: The value of the param if it is found and can be converted
             to a ``bool``. If the param is not found, returns ``None``
-            unless required is ``True``.
+            unless `required` is ``True``.
 
         Raises:
             HTTPBadRequest: A required param is missing from the request, or
@@ -1425,8 +1431,8 @@ class Request(object):
                 val = True
             elif val in FALSE_STRINGS:
                 val = False
-            elif blank_as_true and not val:
-                val = True
+            elif not val:
+                val = blank_as_true
             else:
                 msg = 'The value of the parameter must be "true" or "false".'
                 raise errors.HTTPInvalidParam(msg, name)
@@ -1734,8 +1740,8 @@ class Request(object):
         if body:
             extra_params = parse_query_string(
                 body,
-                keep_blank_qs_values=self.options.keep_blank_qs_values,
-                parse_qs_csv=self.options.auto_parse_qs_csv,
+                keep_blank=self.options.keep_blank_qs_values,
+                csv=self.options.auto_parse_qs_csv,
             )
 
             self._params.update(extra_params)
@@ -1749,8 +1755,8 @@ class RequestOptions(object):
     configuring certain :py:class:`~.Request` behaviors.
 
     Attributes:
-        keep_blank_qs_values (bool): Set to ``True`` to keep query string
-            fields even if they do not have a value (default ``False``).
+        keep_blank_qs_values (bool): Set to ``False`` to ignore query string
+            params that have missing or blank values (default ``True``).
             For comma-separated values, this option also determines
             whether or not empty elements in the parsed list are
             retained.
@@ -1779,14 +1785,15 @@ class RequestOptions(object):
                 encoded according to the standard W3C algorithm (see
                 also http://goo.gl/6rlcux).
 
-        auto_parse_qs_csv: Set to ``False`` to treat commas in a query
-            string value as literal characters, rather than as a comma-
-            separated list (default ``True``). When this option is
-            enabled, the value will be split on any non-percent-encoded
-            commas. Disable this option when encoding lists as multiple
-            occurrences of the same parameter, and when values may be
-            encoded in alternative formats in which the comma character
-            is significant.
+        auto_parse_qs_csv: Set to ``True`` to split query string values on
+            any non-percent-encoded commas (default ``False``). When ``False``,
+            values containing commas are left as-is. In this mode, list items
+            are taken only from multiples of the same parameter name within the
+            query string (i.e. ``/?t=1,2,3&t=4`` becomes ``['1,2,3', '4']``).
+            When `auto_parse_qs_csv` is set to ``True``, the query string value
+            is also split on non-percent-encoded commas and these items
+            are added to the final list (i.e. ``/?t=1,2,3&t=4``
+            becomes ``['1', '2', '3', '4']``).
 
         strip_url_path_trailing_slash: Set to ``True`` in order to
             strip the trailing slash, if present, at the end of the URL
@@ -1819,9 +1826,9 @@ class RequestOptions(object):
     )
 
     def __init__(self):
-        self.keep_blank_qs_values = False
+        self.keep_blank_qs_values = True
         self.auto_parse_form_urlencoded = False
-        self.auto_parse_qs_csv = True
+        self.auto_parse_qs_csv = False
         self.strip_url_path_trailing_slash = False
         self.default_media_type = DEFAULT_MEDIA_TYPE
         self.media_handlers = Handlers()