+++ falcon/testing/client.py
@@ -21,6 +21,7 @@ WSGI callable, without having to stand up a WSGI server.
 import warnings
 import wsgiref.validate
 
+import six
 from six.moves import http_cookies
 
 from falcon.constants import MEDIA_JSON
@@ -231,7 +232,8 @@ class Cookie(object):
 def simulate_request(app, method='GET', path='/', query_string=None,
                      headers=None, body=None, json=None, file_wrapper=None,
                      wsgierrors=None, params=None, params_csv=True,
-                     protocol='http'):
+                     protocol='http', host=helpers.DEFAULT_HOST,
+                     remote_addr=None, extras=None):
         """Simulates a request to a WSGI application.
 
         Performs a request against a WSGI application. Uses
@@ -274,8 +276,14 @@ def simulate_request(app, method='GET', path='/', query_string=None,
                 environ (default: ``None``). This can be used to test
                 high-performance file transmission when `resp.stream` is
                 set to a file-like object.
+            host(str): A string to use for the hostname part of the fully
+                qualified request URL (default: 'falconframework.org')
+            remote_addr (str): A string to use as the remote IP address for the
+                request (default: '127.0.0.1')
             wsgierrors (io): The stream to use as *wsgierrors*
                 (default ``sys.stderr``)
+            extras (dict): Additional CGI variables to add to the WSGI
+                ``environ`` dictionary for the request (default: ``None``)
 
         Returns:
             :py:class:`~.Result`: The result of the request
@@ -298,6 +306,17 @@ def simulate_request(app, method='GET', path='/', query_string=None,
                 'query_string parameter instead.'
             )
 
+        extras = extras or {}
+        if 'REQUEST_METHOD' in extras and extras['REQUEST_METHOD'] != method:
+            # NOTE(vytas): Even given the duct tape nature of overriding
+            # arbitrary environ variables, changing the method can potentially
+            # be very confusing, particularly when using specialized
+            # simulate_get/post/patch etc methods.
+            raise ValueError(
+                'environ extras may not override the request method. Please '
+                'use the method parameter.'
+            )
+
         if query_string is None:
             query_string = to_query_str(
                 params,
@@ -318,8 +337,12 @@ def simulate_request(app, method='GET', path='/', query_string=None,
             headers=headers,
             body=body,
             file_wrapper=file_wrapper,
+            host=host,
+            remote_addr=remote_addr,
             wsgierrors=wsgierrors,
         )
+        if extras:
+            env.update(extras)
 
         srmock = StartResponseMock()
         validator = wsgiref.validate.validator(app)
@@ -366,6 +389,12 @@ def simulate_get(app, path, **kwargs):
             set to a file-like object.
         protocol: The protocol to use for the URL scheme
             (default: 'http')
+        host(str): A string to use for the hostname part of the fully qualified
+            request URL (default: 'falconframework.org')
+        remote_addr (str): A string to use as the remote IP address for the
+            request (default: '127.0.0.1')
+        extras (dict): Additional CGI variables to add to the WSGI ``environ``
+            dictionary for the request (default: ``None``)
     """
     return simulate_request(app, 'GET', path, **kwargs)
 
@@ -400,6 +429,12 @@ def simulate_head(app, path, **kwargs):
             (default: ``None``)
         protocol: The protocol to use for the URL scheme
             (default: 'http')
+        host(str): A string to use for the hostname part of the fully qualified
+            request URL (default: 'falconframework.org')
+        remote_addr (str): A string to use as the remote IP address for the
+            request (default: '127.0.0.1')
+        extras (dict): Additional CGI variables to add to the WSGI ``environ``
+            dictionary for the request (default: ``None``)
     """
     return simulate_request(app, 'HEAD', path, **kwargs)
 
@@ -439,6 +474,12 @@ def simulate_post(app, path, **kwargs):
             `headers`.
         protocol: The protocol to use for the URL scheme
             (default: 'http')
+        host(str): A string to use for the hostname part of the fully qualified
+            request URL (default: 'falconframework.org')
+        remote_addr (str): A string to use as the remote IP address for the
+            request (default: '127.0.0.1')
+        extras (dict): Additional CGI variables to add to the WSGI ``environ``
+            dictionary for the request (default: ``None``)
     """
     return simulate_request(app, 'POST', path, **kwargs)
 
@@ -478,6 +519,12 @@ def simulate_put(app, path, **kwargs):
             `headers`.
         protocol: The protocol to use for the URL scheme
             (default: 'http')
+        host(str): A string to use for the hostname part of the fully qualified
+            request URL (default: 'falconframework.org')
+        remote_addr (str): A string to use as the remote IP address for the
+            request (default: '127.0.0.1')
+        extras (dict): Additional CGI variables to add to the WSGI ``environ``
+            dictionary for the request (default: ``None``)
     """
     return simulate_request(app, 'PUT', path, **kwargs)
 
@@ -509,6 +556,12 @@ def simulate_options(app, path, **kwargs):
             (default: ``None``)
         protocol: The protocol to use for the URL scheme
             (default: 'http')
+        host(str): A string to use for the hostname part of the fully qualified
+            request URL (default: 'falconframework.org')
+        remote_addr (str): A string to use as the remote IP address for the
+            request (default: '127.0.0.1')
+        extras (dict): Additional CGI variables to add to the WSGI ``environ``
+            dictionary for the request (default: ``None``)
     """
     return simulate_request(app, 'OPTIONS', path, **kwargs)
 
@@ -548,6 +601,12 @@ def simulate_patch(app, path, **kwargs):
             `headers`.
         protocol: The protocol to use for the URL scheme
             (default: 'http')
+        host(str): A string to use for the hostname part of the fully qualified
+            request URL (default: 'falconframework.org')
+        remote_addr (str): A string to use as the remote IP address for the
+            request (default: '127.0.0.1')
+        extras (dict): Additional CGI variables to add to the WSGI ``environ``
+            dictionary for the request (default: ``None``)
     """
     return simulate_request(app, 'PATCH', path, **kwargs)
 
@@ -579,6 +638,12 @@ def simulate_delete(app, path, **kwargs):
             (default: ``None``)
         protocol: The protocol to use for the URL scheme
             (default: 'http')
+        host(str): A string to use for the hostname part of the fully qualified
+            request URL (default: 'falconframework.org')
+        remote_addr (str): A string to use as the remote IP address for the
+            request (default: '127.0.0.1')
+        extras (dict): Additional CGI variables to add to the WSGI ``environ``
+            dictionary for the request (default: ``None``)
     """
     return simulate_request(app, 'DELETE', path, **kwargs)
 
@@ -684,5 +749,9 @@ class TestClient(object):
             merged_headers.update(additional_headers)
 
             kwargs['headers'] = merged_headers
+
+        if six.PY2 and 'headers' in kwargs:
+            for key, value in six.iteritems(kwargs['headers']):
+                kwargs['headers'][key] = str(value) if isinstance(value, unicode) else value
 
         return simulate_request(self.app, *args, **kwargs)