+++ tests/test_after_hooks.py
@@ -303,3 +303,81 @@ def test_wrapped_resource_with_hooks_aware_of_resource(client, wrapped_resource_
     result = client.simulate_options('/wrapped_aware')
     assert result.status_code == 200
     assert not result.text
+
+
+class ResourceAwareGameHook(object):
+
+    VALUES = ('rock', 'scissors', 'paper')
+
+    @classmethod
+    def __call__(cls, req, resp, resource):
+        assert resource
+        assert resource.seed in cls.VALUES
+        assert resp.text == 'Responder called.'
+
+        header = resp.get_header('X-Hook-Game')
+        values = header.split(', ') if header else []
+        if values:
+            last = cls.VALUES.index(values[-1])
+            values.append(cls.VALUES[(last + 1) % len(cls.VALUES)])
+        else:
+            values.append(resource.seed)
+        resp.set_header('X-Hook-Game', ', '.join(values))
+
+
+_game_hook = ResourceAwareGameHook()
+
+
+@falcon.after(_game_hook)
+@falcon.after(_game_hook)
+class HandGame(object):
+
+    def __init__(self):
+        self.seed = None
+
+    @falcon.after(_game_hook)
+    def on_put(self, req, resp):
+        self.seed = req.media
+        resp.text = 'Responder called.'
+
+    @falcon.after(_game_hook)
+    def on_get_once(self, req, resp):
+        resp.text = 'Responder called.'
+
+    @falcon.after(_game_hook)
+    @falcon.after(_game_hook)
+    def on_get_twice(self, req, resp):
+        resp.text = 'Responder called.'
+
+    @falcon.after(_game_hook)
+    @falcon.after(_game_hook)
+    @falcon.after(_game_hook)
+    def on_get_thrice(self, req, resp):
+        resp.text = 'Responder called.'
+
+
+@pytest.fixture
+def game_client():
+    app = falcon.API()
+    resource = HandGame()
+
+    app.add_route('/seed', resource)
+    app.add_route('/once', resource, suffix='once')
+    app.add_route('/twice', resource, suffix='twice')
+    app.add_route('/thrice', resource, suffix='thrice')
+
+    return testing.TestClient(app)
+
+
+@pytest.mark.parametrize('seed,uri,expected', [
+    ('paper', '/once', 'paper, rock, scissors'),
+    ('scissors', '/twice', 'scissors, paper, rock, scissors'),
+    ('rock', '/thrice', 'rock, scissors, paper, rock, scissors'),
+    ('paper', '/thrice', 'paper, rock, scissors, paper, rock'),
+])
+def test_after_hooks_on_suffixed_resource(game_client, seed, uri, expected):
+    game_client.simulate_put('/seed', json=seed)
+
+    resp = game_client.simulate_get(uri)
+    assert resp.status_code == 200
+    assert resp.headers['X-Hook-Game'] == expected