+++ tests/test_response.py
@@ -1,3 +1,5 @@
+import pytest
+
 import falcon
 from falcon import MEDIA_TEXT
 
@@ -11,3 +13,36 @@ def test_response_set_content_type_set():
 def test_response_set_content_type_not_set():
     resp = falcon.Response()
     assert 'content-type' not in resp._headers
+
+
+def test_response_get_headers():
+    resp = falcon.Response()
+    resp.append_header('x-things1', 'thing-1')
+    resp.append_header('x-things2', 'thing-2')
+    resp.append_header('X-Things3', 'Thing-3')
+
+    resp.set_cookie('Chocolate', 'Chip')
+
+    headers = resp.headers
+    assert headers['x-things1'] == 'thing-1'
+    assert headers['x-things2'] == 'thing-2'
+    assert headers['x-things3'] == 'Thing-3'
+
+    assert 'set-cookie' not in headers
+
+
+def test_response_attempt_to_set_read_only_headers():
+    resp = falcon.Response()
+
+    resp.append_header('x-things1', 'thing-1')
+    resp.append_header('x-things2', 'thing-2')
+    resp.append_header('x-things3', 'thing-3a')
+    resp.append_header('X-Things3', 'thing-3b')
+
+    with pytest.raises(AttributeError):
+        resp.headers = {'x-things4': 'thing-4'}
+
+    headers = resp.headers
+    assert headers['x-things1'] == 'thing-1'
+    assert headers['x-things2'] == 'thing-2'
+    assert headers['x-things3'] == 'thing-3a, thing-3b'